backend/requirements.txt
# ==================================
# Godot-Minds Backend Dependencies
# ==================================

# Core Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0

# API & Async
aiohttp==3.9.1
httpx==0.26.0
python-multipart==0.0.6
aiofiles==23.2.1

# Configuration & Environment
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# AI Providers
anthropic==0.7.8
openai==1.10.0
google-generativeai==0.3.2

# Vector Database (RAG)
chromadb==0.4.22

# Git Integration
gitpython==3.1.41

# File Watching
watchdog==4.0.0

# Data Processing
numpy==1.26.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
black==24.1.1
ruff==0.1.15
mypy==1.8.0

# Utilities
python-dateutil==2.8.2
